+++ src/reduxStore/selectors/index.ts
@@ -1,8 +1,8 @@
 export * from './accountInfoSelectors';
-export * from './networkConfigSelectors';
 export * from './loginInfoSelectors';
-
-export * from './transactionsSelectors';
-export * from './transactionsInfoSelectors';
 export * from './modalsSelectors';
+export * from './networkConfigSelectors';
+export * from './signedMessageInfoSelectors';
 export * from './toastsSelectors';
+export * from './transactionsInfoSelectors';
+export * from './transactionsSelectors';