+++ src/types/signedMessage.types.ts
@@ -1,0 +1,5 @@
+export enum SignedMessageQueryParamsEnum {
+  signature = 'signature',
+  sessionId = 'sessionId',
+  status = 'status'
+}